+++ src/pcloud/pcloudfs.py
@@ -10,8 +10,7 @@ from pcloud.api import O_CREAT
 
 
 class PCloudFile(BytesIO):
-    """ A file representation for pCloud files
-    """
+    """A file representation for pCloud files"""
 
     def __init__(self, pcloud, path, mode):
         self.pcloud = pcloud
@@ -142,8 +141,7 @@ class PCloudFS(FS):
         self.check()
         result = self.pcloud.createfolder(path=path)
         if result["result"] == 2004:
-            raise errors.DirectoryExists(
-                'Directory "{0}" already exists'.format(path))
+            raise errors.DirectoryExists('Directory "{0}" already exists'.format(path))
         elif result["result"] != 0:
             raise errors.CreateFailed(
                 'Create of directory "{0}" failed with "{1}"'.format(