+++ src/pcloud/tests/conftest.py
@@ -8,8 +8,7 @@ PORT = 5000
 
 @pytest.fixture(scope="session")
 def start_mock_server():
-    """ Start a mock server on port 5000
-    """
+    """Start a mock server on port 5000"""
     httpd = MockServer(("", PORT), MockHandler)
     httpd_thread = Thread(target=httpd.serve_forever, daemon=True)
     httpd_thread.start()