+++ CHANGES.rst
@@ -17,6 +17,18 @@ Changelog
 - Add new stats endpoint
   [AgusRumayor]
 
+- Add methods for archiving
+  [olokelo]
+
+- Add token expire parameter
+  [olekelo]
+
+- Start implementing trash methods
+  [qo4on, tomgross]
+
+- Add support for alternate endpoints
+  [tomgross]
+
 1.0a8 (2020-02-21)
 ------------------